Add QE verification step after POST+built
After a bootimage bug lands in an RHCOS build, ask QE to verify the bug
and to set the Verified field to Tested.  Add icon for "awaiting testing"
and show the corresponding bug text unitalicized.

Don't change the conditions for moving a bootimage bug to MODIFIED, since
the fix will still have landed in a bootimage whether it's been verified
or not.